Remove setCaretToTextEnd from PostComment
This feature was quite annoying since it worked on every focusing,
so if you adding text in the middle of the comment and blurring
textarea for some reason, after focusing back it sets caret to the
end again.

Automatic moving it to the text end on INITIAL rendering of textarea
with autoFocus still works though (at least in Chrome and Firefox),
that's sweet!