Use cookie domain
This fixes the issue with inability to sign out at
"gamma.freefeed.net" when there's still the cookie from old Ember
client with domain ".freefeed.net" left in the browser.

To solve this, one needs to apply this change and set getCookieDomain
in config.js to `() => '.freefeed.net'`.

It's implemented as a function to enable support for dynamic resolution
(e.g., based on window.location).