[components-rewiring] Untangle PostComments, part 3/3
This one also fixes the bug with editing comments:
"if during the editing of the last comment, a new (realtime) comment
arrives to the thread, the app resets ongoing changes in the textarea".

It was caused by comment moving from "last" to "middle" with
corresponding removing from the DOM and inserting it back with
different key. No more!