Add throbber for "X more comments" on Feed->Post transition
When the post is loaded with collapsed comments (feed) and user
goes to the post's page, the post shows up immediately, but with
comments still collapsed. We should activate `isLoadingComments`
to display a throbber on "X more comments", otherwise user on slow
network connection would want to expand them manually.