Rewrite toggleSidebar() to change DOM directly
Instead of keeping sidebarViewState and re-render Mobile Sidebar every
time it shows up, we remove sidebarViewState from Redux completely and
update class names directly in DOM. It means the state of Sidebar is
quite ephemeral (exists in DOM only), but the performance boost is
worth it.

This makes Mobile Sidebar 3.5x faster: 14 fps.