Fix relative date if server time is in future
We won't have weird "in a few seconds" for brand new entries, as it was
with the old implementation, but still there might be strange countdown
if server time is a little ahead. Like, "2 mins ago", then "1 min ago",
then "just now" and then it will begin increasing back.

This change fixes it, it will be showing "just now" until client catch
up server time.