Highlight and scroll to comment when arrows (^^^) clicked
We piggyback on START_HIGHLIGHTING_COMMENTS action to notify middleware
and redirect to #comment-abcd. Somewhat "hacky", but whatever.

Also, in this change we bring back componentDidUpdate(), to enable
scrolling between comments within the post. That method was removed in
commit bd1113d on 2017-05-17. This new version, however, have a special
condition that neutralizes the side effects we had in the old version.