clikes: Use manual trigger for tooltips
We'll need it to support liking/unliking on the second click (see the
following commits).

We use react-tippy (based on tippyjs) which hides the tooltip on
clicking outside of it. There is exception for the trigger element, but
it only works when trigger event is "click", not "manual". The solution
is to use customized version of react-tippy that supports it.