Fix post bump for present hidden posts
Before:

A new comment to a hidden post that is present on the page
caused that post to show up in the top (as freshly hidden entry,
i.e. "Entry hidden - undo.").

After:

A new comment is added to the existing post in the "Hidden entries"
section in the bottom of the page. No unexpected entries with
"Entry hidden - undo" text anymore.