Fix hiding/un-hiding non-present posts brought by realtime
Before:

JS errors in tab A (with opened home, 1st page), when the user in
tab B (with page 2 opened) hides/un-hides posts.

They happen because it tries to handle the posts that are not in
state.posts.

After:

No changes in tab A are necessary and none are made (and no JS errors,
either).