group-moderation: Fix post re-rendering (memoize properly)
It seems global post re-rendering has been happening for long time,
because state.feeds and state.users were in the "dependencies" (and
earlier, it was state.subscribers and state.subscriptions).

Not anymore. Must be important optimization?

BTW, avoiding unnecessary posts' re-rendering fixes a bug. There was
an issue with viewing zoomed-in images: every realtime update caused
the image's zoom to reset (because the post of that image got
re-rendered). It was especially annoying on mobile. Not anymore.