Fix "direct to nobody"
When the recipient of your direct message removes themselves from the
thread, your post might end up without any recipients (but it is still
a direct message, not a post to your personal feed).

This fixes this case, so it doesn't break the app, and so it displays
the proper visibility icon.